+++ src/main/java/dankook/capstone/petalk/dto/response/UploadVideoResponse.java
@@ -1,12 +1,15 @@
-package dankook.capstone.petalk.dto;
+package dankook.capstone.petalk.dto.response;
 
 import dankook.capstone.petalk.domain.Emotion;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.io.File;
+
 @Data
 @AllArgsConstructor
-public class VideoEmotionResponse {
+public class UploadVideoResponse {
     private Long id;
+    private File video;
     private Emotion emotion;
 }
