+++ src/main/java/dankook/capstone/petalk/dto/response/VideoEmotionResponse.java
@@ -1,0 +1,12 @@
+package dankook.capstone.petalk.dto.response;
+
+import dankook.capstone.petalk.domain.Emotion;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class VideoEmotionResponse {
+    private Long id;
+    private Emotion emotion;
+}
