+++ src/main/java/dankook/capstone/petalk/repository/VideoRepository.java
@@ -1,7 +1,7 @@
 package dankook.capstone.petalk.repository;
 
 import dankook.capstone.petalk.domain.Video;
-import dankook.capstone.petalk.dto.VideoDto;
+import dankook.capstone.petalk.dto.response.VideoDto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
